backend/requirements.txt
Flask==2.3.3
{{ ... }}
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.5.0
python-dotenv==1.0.0
PyJWT==2.8.0
# bcrypt==4.0.1  # Optional - using werkzeug for password hashing instead
Pillow==10.0.0
numpy==1.24.3
tensorflow==2.13.0
keras==2.13.1
requests==2.31.0
python-multipart==0.0.6
Werkzeug==2.3.7
beautifulsoup4==4.12.2

# Optional: For ML model training
# matplotlib==3.7.2
# tensorflow-datasets==4.9.2
